requirements.txt
selenium~=4.28.1
python-dotenv~=1.0.1